Extract editor content update logic to custom hook
